+++ sigul-pesign-bridge/Cargo.toml
@@ -67,6 +67,10 @@ features = ["derive"]
 version = "1.6"
 features = ["v7"]
 
+[dependencies.zerocopy]
+version = "0.8"
+features = ["std", "derive"]
+
 [dev-dependencies]
 tracing-test = "0.2"
 assert_cmd = "2"
