+++ sigul-pesign-bridge/src/lib.rs
@@ -6,6 +6,7 @@
 #[doc(hidden)]
 pub mod cli;
 pub mod config;
+pub(crate) mod pesign;
 mod service;
 
 #[doc(hidden)]
